update deleteImage helper

+++ server/src/helper/deleteImage.js
@@ -1,15 +1,19 @@
 const fs = require("fs").promises;
 const logger = require("./winstonLogger");
 
-const deleteImage = async (imagePath) => {
-    try {
-        await fs.access(imagePath);
-        await fs.unlink(imagePath);
-        logger.info('Image deleted successfully')
-    } catch (error) {
-        logger.error('Image does not exist');
-        throw error;
+const deleteImage = async (imagePath, defaultImagePath = "") => {
+  try {
+    if (imagePath !== defaultImagePath) {
+      await fs.access(imagePath);
+      await fs.unlink(imagePath);
+      logger.info("Image deleted successfully");
+    } else {
+      logger.info("Default Image will not be deleted");
     }
-}
+  } catch (error) {
+    logger.error("Image does not exist");
+    throw error;
+  }
+};
 
 module.exports = deleteImage;
